+++ tests/test_multiprocessing_serverprocess.py
@@ -6,6 +6,8 @@ from multiprocessing import active_children
 from traceback import print_exc
 
 import pytest
+from zpz.logging import config_logger, unuse_console_handler
+
 from mpservice.multiprocessing import Process, Queue, SpawnProcess
 from mpservice.multiprocessing.server_process import (
     MemoryBlock,
@@ -15,7 +17,6 @@ from mpservice.multiprocessing.server_process import (
     managed_memoryblock,
 )
 from mpservice.threading import Thread
-from zpz.logging import config_logger, unuse_console_handler
 
 unuse_console_handler()
 config_logger()